for fixed2.

+++ Scripts/for/for19.py
@@ -1,7 +1,7 @@
 # for19 Прохоренков Егор ИСП-211.
-n = float(input())
+n = int(input())
 
-factor = 1
+factor = 1.0
 
 for i in range(1, n + 1):
     factor *= i
